CMake update and MSVS (#6)
* basic compilation

* added the images

* translations and installation

* added win32 and D-Bus specific stuff

* adaptions for MS Visual Studio

- _USE_MATH_DEFINES is still necessary
- MSVS looks up for mangled symbold, i.e. Settings != QSettings, etc.

+++ src/treeitem.cpp
@@ -12,7 +12,7 @@
 
 using namespace std;
 
-extern uint itemLastID;
+extern ulong itemLastID;
 extern FlagRowMaster *standardFlagsMaster;
 extern FlagRowMaster *systemFlagsMaster;
 extern FlagRowMaster *userFlagsMaster;
